Wayland: Split WaylandThread implementation into multiple files
Wayland is an asynchronous object-oriented API. This means that most
code in WaylandThread is made of event handlers for each kind of
object. Over time it all became quite unwieldy.

This patch moves the implementation for all the handlers into separate
files, one for each protocol. The core protocol got further split into
files as it's quite chunky.

Note that this relates only to the _implementation_ of the
`WaylandThread` class (into which the handlers are currently located)
and does no further attempt at splitting everything up (e.g. our custom
object state structs), since that would require a lot more work.

Personally I think that this is definitely clearer, given how simple
this change is, and that potentially we might go even further over time.

Perhaps we could use a "wayland" namespace for all the state objects and
get rid of them from the `WaylandThread` class? Just thinking out loud
:P
