+++ platform/linuxbsd/wayland/thread/core/keyboard.cpp
@@ -1,0 +1,145 @@
+/**************************************************************************/
+/*  keyboard.cpp                                                          */
+/**************************************************************************/
+/*                         This file is part of:                          */
+/*                             GODOT ENGINE                               */
+/*                        https://godotengine.org                         */
+/**************************************************************************/
+/* Copyright (c) 2014-present Godot Engine contributors (see AUTHORS.md). */
+/* Copyright (c) 2007-2014 Juan Linietsky, Ariel Manzur.                  */
+/*                                                                        */
+/* Permission is hereby granted, free of charge, to any person obtaining  */
+/* a copy of this software and associated documentation files (the        */
+/* "Software"), to deal in the Software without restriction, including    */
+/* without limitation the rights to use, copy, modify, merge, publish,    */
+/* distribute, sublicense, and/or sell copies of the Software, and to     */
+/* permit persons to whom the Software is furnished to do so, subject to  */
+/* the following conditions:                                              */
+/*                                                                        */
+/* The above copyright notice and this permission notice shall be         */
+/* included in all copies or substantial portions of the Software.        */
+/*                                                                        */
+/* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,        */
+/* EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF     */
+/* MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. */
+/* IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY   */
+/* CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,   */
+/* TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE      */
+/* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
+/**************************************************************************/
+
+#include "wayland/wayland_thread.h"
+
+#include <sys/mman.h>
+
+void WaylandThread::_wl_keyboard_on_keymap(void *data, struct wl_keyboard *wl_keyboard, uint32_t format, int32_t fd, uint32_t size) {
+	ERR_FAIL_COND_MSG(format != WL_KEYBOARD_KEYMAP_FORMAT_XKB_V1, "Unsupported keymap format announced from the Wayland compositor.");
+
+	SeatState *ss = (SeatState *)data;
+	ERR_FAIL_NULL(ss);
+
+	if (ss->keymap_buffer) {
+		// We have already a mapped buffer, so we unmap it. There's no need to reset
+		// its pointer or size, as we're gonna set them below.
+		munmap((void *)ss->keymap_buffer, ss->keymap_buffer_size);
+		ss->keymap_buffer = nullptr;
+	}
+
+	ss->keymap_buffer = (const char *)mmap(nullptr, size, PROT_READ, MAP_PRIVATE, fd, 0);
+	ss->keymap_buffer_size = size;
+
+	xkb_keymap_unref(ss->xkb_keymap);
+	ss->xkb_keymap = xkb_keymap_new_from_string(ss->xkb_context, ss->keymap_buffer,
+			XKB_KEYMAP_FORMAT_TEXT_V1, XKB_KEYMAP_COMPILE_NO_FLAGS);
+
+	xkb_state_unref(ss->xkb_state);
+	ss->xkb_state = xkb_state_new(ss->xkb_keymap);
+}
+
+void WaylandThread::_wl_keyboard_on_enter(void *data, struct wl_keyboard *wl_keyboard, uint32_t serial, struct wl_surface *surface, struct wl_array *keys) {
+	SeatState *ss = (SeatState *)data;
+	ERR_FAIL_NULL(ss);
+
+	WaylandThread *wayland_thread = ss->wayland_thread;
+	ERR_FAIL_NULL(wayland_thread);
+
+	wayland_thread->_set_current_seat(ss->wl_seat);
+
+	Ref<WindowEventMessage> msg;
+	msg.instantiate();
+	msg->event = DisplayServer::WINDOW_EVENT_FOCUS_IN;
+	wayland_thread->push_message(msg);
+}
+
+void WaylandThread::_wl_keyboard_on_leave(void *data, struct wl_keyboard *wl_keyboard, uint32_t serial, struct wl_surface *surface) {
+	SeatState *ss = (SeatState *)data;
+	ERR_FAIL_NULL(ss);
+
+	WaylandThread *wayland_thread = ss->wayland_thread;
+	ERR_FAIL_NULL(wayland_thread);
+
+	ss->repeating_keycode = XKB_KEYCODE_INVALID;
+
+	Ref<WindowEventMessage> msg;
+	msg.instantiate();
+	msg->event = DisplayServer::WINDOW_EVENT_FOCUS_OUT;
+	wayland_thread->push_message(msg);
+}
+
+void WaylandThread::_wl_keyboard_on_key(void *data, struct wl_keyboard *wl_keyboard, uint32_t serial, uint32_t time, uint32_t key, uint32_t state) {
+	SeatState *ss = (SeatState *)data;
+	ERR_FAIL_NULL(ss);
+
+	WaylandThread *wayland_thread = ss->wayland_thread;
+	ERR_FAIL_NULL(wayland_thread);
+
+	// We have to add 8 to the scancode to get an XKB-compatible keycode.
+	xkb_keycode_t xkb_keycode = key + 8;
+
+	bool pressed = state & WL_KEYBOARD_KEY_STATE_PRESSED;
+
+	if (pressed) {
+		if (xkb_keymap_key_repeats(ss->xkb_keymap, xkb_keycode)) {
+			ss->last_repeat_start_msec = OS::get_singleton()->get_ticks_msec();
+			ss->repeating_keycode = xkb_keycode;
+		}
+
+		ss->last_key_pressed_serial = serial;
+	} else if (ss->repeating_keycode == xkb_keycode) {
+		ss->repeating_keycode = XKB_KEYCODE_INVALID;
+	}
+
+	Ref<InputEventKey> k;
+	k.instantiate();
+
+	if (!_seat_state_configure_key_event(*ss, k, xkb_keycode, pressed)) {
+		return;
+	}
+
+	Ref<InputEventMessage> msg;
+	msg.instantiate();
+	msg->event = k;
+	wayland_thread->push_message(msg);
+}
+
+void WaylandThread::_wl_keyboard_on_modifiers(void *data, struct wl_keyboard *wl_keyboard, uint32_t serial, uint32_t mods_depressed, uint32_t mods_latched, uint32_t mods_locked, uint32_t group) {
+	SeatState *ss = (SeatState *)data;
+	ERR_FAIL_NULL(ss);
+
+	xkb_state_update_mask(ss->xkb_state, mods_depressed, mods_latched, mods_locked, ss->current_layout_index, ss->current_layout_index, group);
+
+	ss->shift_pressed = xkb_state_mod_name_is_active(ss->xkb_state, XKB_MOD_NAME_SHIFT, XKB_STATE_MODS_DEPRESSED);
+	ss->ctrl_pressed = xkb_state_mod_name_is_active(ss->xkb_state, XKB_MOD_NAME_CTRL, XKB_STATE_MODS_DEPRESSED);
+	ss->alt_pressed = xkb_state_mod_name_is_active(ss->xkb_state, XKB_MOD_NAME_ALT, XKB_STATE_MODS_DEPRESSED);
+	ss->meta_pressed = xkb_state_mod_name_is_active(ss->xkb_state, XKB_MOD_NAME_LOGO, XKB_STATE_MODS_DEPRESSED);
+
+	ss->current_layout_index = group;
+}
+
+void WaylandThread::_wl_keyboard_on_repeat_info(void *data, struct wl_keyboard *wl_keyboard, int32_t rate, int32_t delay) {
+	SeatState *ss = (SeatState *)data;
+	ERR_FAIL_NULL(ss);
+
+	ss->repeat_key_delay_msec = 1000 / rate;
+	ss->repeat_start_delay_msec = delay;
+}
