+++ platform/linuxbsd/wayland/thread/primary-selection.cpp
@@ -1,0 +1,105 @@
+/**************************************************************************/
+/*  primary-selection.cpp                                                 */
+/**************************************************************************/
+/*                         This file is part of:                          */
+/*                             GODOT ENGINE                               */
+/*                        https://godotengine.org                         */
+/**************************************************************************/
+/* Copyright (c) 2014-present Godot Engine contributors (see AUTHORS.md). */
+/* Copyright (c) 2007-2014 Juan Linietsky, Ariel Manzur.                  */
+/*                                                                        */
+/* Permission is hereby granted, free of charge, to any person obtaining  */
+/* a copy of this software and associated documentation files (the        */
+/* "Software"), to deal in the Software without restriction, including    */
+/* without limitation the rights to use, copy, modify, merge, publish,    */
+/* distribute, sublicense, and/or sell copies of the Software, and to     */
+/* permit persons to whom the Software is furnished to do so, subject to  */
+/* the following conditions:                                              */
+/*                                                                        */
+/* The above copyright notice and this permission notice shall be         */
+/* included in all copies or substantial portions of the Software.        */
+/*                                                                        */
+/* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,        */
+/* EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF     */
+/* MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. */
+/* IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY   */
+/* CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,   */
+/* TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE      */
+/* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
+/**************************************************************************/
+
+#include "wayland/wayland_thread.h"
+
+#include <unistd.h>
+
+// NOTE: Don't forget to `memfree` the offer's state.
+void WaylandThread::_wp_primary_selection_device_on_data_offer(void *data, struct zwp_primary_selection_device_v1 *wp_primary_selection_device_v1, struct zwp_primary_selection_offer_v1 *offer) {
+	wl_proxy_tag_godot((struct wl_proxy *)offer);
+	zwp_primary_selection_offer_v1_add_listener(offer, &wp_primary_selection_offer_listener, memnew(OfferState));
+}
+
+void WaylandThread::_wp_primary_selection_device_on_selection(void *data, struct zwp_primary_selection_device_v1 *wp_primary_selection_device_v1, struct zwp_primary_selection_offer_v1 *id) {
+	SeatState *ss = (SeatState *)data;
+	ERR_FAIL_NULL(ss);
+
+	if (ss->wp_primary_selection_offer) {
+		memfree(wp_primary_selection_offer_get_offer_state(ss->wp_primary_selection_offer));
+		zwp_primary_selection_offer_v1_destroy(ss->wp_primary_selection_offer);
+	}
+
+	ss->wp_primary_selection_offer = id;
+}
+
+void WaylandThread::_wp_primary_selection_offer_on_offer(void *data, struct zwp_primary_selection_offer_v1 *wp_primary_selection_offer_v1, const char *mime_type) {
+	OfferState *os = (OfferState *)data;
+	ERR_FAIL_NULL(os);
+
+	if (os) {
+		os->mime_types.insert(String::utf8(mime_type));
+	}
+}
+
+void WaylandThread::_wp_primary_selection_source_on_send(void *data, struct zwp_primary_selection_source_v1 *wp_primary_selection_source_v1, const char *mime_type, int32_t fd) {
+	SeatState *ss = (SeatState *)data;
+	ERR_FAIL_NULL(ss);
+
+	Vector<uint8_t> *data_to_send = nullptr;
+
+	if (wp_primary_selection_source_v1 == ss->wp_primary_selection_source) {
+		data_to_send = &ss->primary_data;
+		DEBUG_LOG_WAYLAND_THREAD("Clipboard: requested primary selection.");
+	}
+
+	if (data_to_send) {
+		ssize_t written_bytes = 0;
+
+		if (strcmp(mime_type, "text/plain") == 0) {
+			written_bytes = write(fd, data_to_send->ptr(), data_to_send->size());
+		}
+
+		if (written_bytes > 0) {
+			DEBUG_LOG_WAYLAND_THREAD(vformat("Clipboard: sent %d bytes.", written_bytes));
+		} else if (written_bytes == 0) {
+			DEBUG_LOG_WAYLAND_THREAD("Clipboard: no bytes sent.");
+		} else {
+			ERR_PRINT(vformat("Clipboard: write error %d.", errno));
+		}
+	}
+
+	close(fd);
+}
+
+void WaylandThread::_wp_primary_selection_source_on_cancelled(void *data, struct zwp_primary_selection_source_v1 *wp_primary_selection_source_v1) {
+	SeatState *ss = (SeatState *)data;
+	ERR_FAIL_NULL(ss);
+
+	if (wp_primary_selection_source_v1 == ss->wp_primary_selection_source) {
+		zwp_primary_selection_source_v1_destroy(ss->wp_primary_selection_source);
+		ss->wp_primary_selection_source = nullptr;
+
+		ss->primary_data.clear();
+
+		DEBUG_LOG_WAYLAND_THREAD("Clipboard: primary selection set by another program.");
+		return;
+	}
+}
